feat(launchpad-grc20): transform launchpadButton into a generic LargeBoxButton & import it into ERC20 launchpad screen
